requirements.txt
torch
torchvision
numpy
pandas
opencv-python
scikit-learn
Pillow
ollama
streamlit
streamlit-lottie
requests
